### 2020-06-07 #2297 @whedon
[PRE REVIEW]: Implicit measures with reproducible results: The implicitMeasures package
**Submitting author:** @OttaviaE (<a href="http://orcid.org/0000-0001-8552-568X">Ottavia M. Epifania</a>)
**Repository:** <a href="https://github.com/OttaviaE/implicitMeasures" target ="_blank">https://github.com/OttaviaE/implicitMeasures</a>
**Version:** v0.1.3
