### 2020-06-07 #2297 @kyleniemeyer
@openjournals/dev any ideas why whedon can't find the paper here? Could it have anything to do with the
```yaml
output:
  html_document:
       keep_md: true
```
in the paper header?
