### 2020-06-08 #2297 @kyleniemeyer
Hi @OttaviaE, sure, you may suggest someone, but who we ultimately ask/assign to edit also depends on editor availability and preference
