### 2020-06-15 #2297 @oliviaguest
I have no idea what is wrong with building the PDF? I assume this is a bug? @kyleniemeyer and @openjournals/dev are aware so I'll leave it for now, unless you have ideas, @OttaviaE?

But anyway while we wait, can you fix the Travis build badge, please here: https://github.com/OttaviaE/implicitMeasures/blob/master/README.md?
