+++ README.md
@@ -4,8 +4,8 @@
 # implicitMeasures
 
 R package for computing different Implicit Measures scores
-<!-- badges: start --> [Travis build
-status](https://travis-ci.org/OttaviaE/implicitMeasures.svg?branch=master)\](<https://travis-ci.org/OttaviaE/implicitMeasures>)
+<!-- badges: start --> 
+[Travis build status](https://travis-ci.org/OttaviaE/implicitMeasures.svg?branch=master)\](<https://travis-ci.org/OttaviaE/implicitMeasures>)
 <!-- badges: end -->
 
 ## Aim and Overview
