+++ README.md
@@ -5,7 +5,7 @@
 
 R package for computing different Implicit Measures scores
 <!-- badges: start --> 
-[Travis build status](https://travis-ci.org/OttaviaE/implicitMeasures.svg?branch=master)\](<https://travis-ci.org/OttaviaE/implicitMeasures>)
+[![Build Status](https://travis-ci.com/OttaviaE/implicitMeasures.svg?branch=master)](https://travis-ci.com/OttaviaE/implicitMeasures)
 <!-- badges: end -->
 
 ## Aim and Overview
