+++ README.Rmd
@@ -16,9 +16,9 @@ knitr::opts_chunk$set(
 # implicitMeasures
 
 R package for computing different Implicit Measures scores
-  <!-- badges: start -->
-  [Travis build status](https://travis-ci.org/OttaviaE/implicitMeasures.svg?branch=master)](https://travis-ci.org/OttaviaE/implicitMeasures)
-  <!-- badges: end -->
+
+[![Build Status](https://travis-ci.org/OttaviaE/implicitMeasures.svg?branch=master)](https://travis-ci.org/OttaviaE/implicitMeasures)
+
   
 ## Aim and Overview
 
