### 2020-06-15 #2297 @danielskatz
The repository was incorrectly specified as https://github.com/OttaviaE/implicitMeasures/tree/master/inst rather than https://github.com/OttaviaE/implicitMeasures

I've changed it in this issue - but I'm not sure if it needs to be changed in database as well. cc @openjournals/dev
