### 2020-06-15 #2297 @danielskatz
looks like that fixed it.

Also, @oliviaguest, I removed the waitlisted label, which you should do when you grab a submission of the list to edit in the future.
