### 2020-06-15 #2297 @OttaviaE
> PS: See: [OttaviaE/implicitMeasures#1](https://github.com/OttaviaE/implicitMeasures/pull/1)

Hey @oliviaguest I think I solved the Travis issue, glad to see that also the pdf generation went well, thanks @danielskatz
