### 2020-06-16 #2297 @OttaviaE
Hey @oliviaguest thank you for your work! Yeah I know the struggle, I had problems finding reviewers also when I submitted a paper for a shiny app always related to the IAT... Anyways, I think I found some names, should I mention directly them here?
Thanks again =)
