### 2020-06-16 #2297 @OttaviaE
Hey @mikerspencer and @mmrabe, @oliviaguest  and I were wondering whether you wanted to review my R package for computing implicit measures scores and its related paper....? Thanks!
