### 2020-06-16 #2297 @oliviaguest
Ah, sorry. I should have said not to tag them (see top):

> @OttaviaE if you have any suggestions for potential reviewers then please mention them here in this thread (without tagging them with an @). In addition, this list of people have already agreed to review for JOSS and may be suitable for this submission (please start at the bottom of the list).

I apologise — my fault for not being clear and not reminding you. No worries though. I am sure they are understanding and aware of how to unsubscribe. Sorry.

If you have any other names that are not on that list, also feel free to mention them. I think it might be good to find IAT experts who might be less (explicitly) expert in R per se. Since such people might be your users too, right?
