### 2020-06-16 #2297 @danielskatz
Just to make it clear - Choosing, inviting, and assigning reviewers is part of the job of the editor, though it's typical that she asks the author for suggestions, which may or may not be used. The authors can suggest potential reviewers, but should not ask them to review.
