### 2020-06-16 #2297 @oliviaguest
My fault @danielskatz because I didn't realise/think of the ambiguity between tag and mention — I was hoping just for emails/[user]names.
