### 2020-06-16 #2297 @OttaviaE
@oliviaguest I was thinking maybe we can ask again to one of the reviewer that reviewed my Shiny app... worst case sceanario he can mention someone pretty familiar with the IAT & R. Sorry again
