### 2020-06-16 #2297 @oliviaguest
> Hey @mikerspencer and @mmrabe, @oliviaguest and I were wondering whether you wanted to review my R package for computing implicit measures scores and its related paper....? Thanks!

FWIW if either or both of them want to review this, I think that's a good idea BTW. So I am happy to hear from them if they do check this thread. Otherwise, I will email them soon — I have already sent quite a few emails today for this, so maybe people will get back to us.

> @oliviaguest I was thinking maybe we can ask again to one of the reviewer that reviewed my Shiny app... worst case sceanario he can mention someone pretty familiar with the IAT & R. Sorry again

Drop their details here (ideally their website?) please! Good idea!
