### 2020-06-16 #2297 @OttaviaE
@oliviaguest okay just to clear the air and prevent me from doing any other silly mistakes... can I just write it down his e-mail adrees here? he is in the list of the JOSS reviewers
