### 2020-06-16 #2297 @oliviaguest
@OttaviaE you can write anything except actually @-tagging, get me? See this as "safe" example: https://github.com/openjournals/joss-reviews/issues/2269#issuecomment-644718399
