### 2020-06-16 #2297 @OttaviaE
@oliviaguest got you, thanks. tomfaulkenberry and benmarwick were the reviewers for my shiny app, they looked pretty familiar with both R and the IAT. Maybe they are available or they can suggest someone.
