### 2020-06-24 OttaviaE/implicitMeasures#2 @eirini-zormpa
Warnings when using plotting functions
I get the following warnings when using the `d_plot` function:
```
> d_plot(dscore)
Warning messages:
1: Use of `data$d_cres` is discouraged. Use `d_cres` instead.
2: Use of `data$d` is discouraged. Use `d` instead.
```
There are similar warnings for the `d_dist`, `multi_scores[[2]]`, and `multi_dsciat()` functions. This is not a big issue though, the plots are generated without problems!

#openjournals/joss-reviews#2297
