### 2020-06-24 OttaviaE/implicitMeasures#3 @eirini-zormpa
Question about data structure
I have never used IAT measures, so this may be a naive question!

The `clean_IAT()` function seems to me to require a fairly rigid data structure to work. Would this function be easy to use with a different dataset? That is, are data from IAT experiments typically already in that structure or do they require any preprocessing? If they do require preprocessing, I think it would be helpful to include more information on how users need to prepare their data before using this function.

#openjournals/joss-reviews#2297
