### 2020-06-25 OttaviaE/implicitMeasures#2 @OttaviaE
Hey @eirini-zormpa, yep, I know about that warning. To fix that, i should get rid of `data`, but when I do that and check the package with devtoos::check() I get a warning mesagge becasue the funtcion can't find the object d_cres (and this is true for all the the other graphical functions), even though the graphical representation are reproduced without problems. Since I submitted the package to CRAN, I needed it to have no warnings or notes, that's why I decided to keep it the way it is now!
