### 2020-06-25 OttaviaE/implicitMeasures#3 @OttaviaE
Actually, the point of the clean_iat() function(as well as the clean_iat() function) is to be as flexible as possible. IAT data are usaully collected by using the sofatware Inquisit, whie results in a data set simialr to that used as an example. Anyways, regardless of the specific software used for the administration, the IAT usually results in a dataset with respondents IDs, their RTs, their accuracy responses, and the lables of the critical blocks. Once the users have this infromation, they can use the function with no further problems, by just using the specific lables for each arument. To answer your question, users should not be reuqired to do any preprocessing for using the function that is supposed to to all the preprocessing work for them.
