### 2020-06-25 #2394 @whedon
[REVIEW]: Implicit measures with reproducible results: The implicitMeasures package
**Submitting author:** @OttaviaE (<a href="http://orcid.org/0000-0001-8552-568X">Ottavia M. Epifania</a>)
**Repository:** <a href="https://github.com/OttaviaE/implicitMeasures/" target ="_blank">https://github.com/OttaviaE/implicitMeasures/</a>
**Version:** v0.2.0
