### 2020-06-26 #2394 @wjakethompson
Hi @oliviaguest!

This is very minor, but the `repository_url` link in the checklist is taking me to [`implicitMeasures/inst/`](https://github.com/OttaviaE/implicitMeasures/tree/master/inst) instead of the main repo page (i.e., [`implicitMeasures`](https://github.com/OttaviaE/implicitMeasures)). This doesn't bother me at all, but thought I'd drop a note in case it was indicative of some mistake in the metadata that may cause problems down the line.
