### 2020-06-26 #2394 @oliviaguest
Ah, yes! @danielskatz noticed this (and even called this) previously. I'll tag @openjournals/dev see what they say and how it can be fixed. I manually edited it above in one place (see history of OP), but that might not be enough.
