### 2020-06-26 #2394 @xuanxu
If the url is corrected in the issue body that should be enough. The paper compiler extracts the info from there.
