### 2020-06-26 #2394 @OttaviaE
Hi @wjakethompson and @oliviaguest the paper is in implicitMeasures/inst/ instead of the main repo implicitMeasures beacuse the package and its repository is available on CRAN and CRAN dosen't allow anything else to be on the leel but a folder named "inst"
