### 2020-06-26 #2394 @xuanxu
Yeah, the paper files can be located anywhere in the repo, whedon will look for them in all subfolders.
