### 2020-06-26 #2394 @danielskatz
I fixed the 2 links in the checklist items for "repository URL"
