### 2020-06-30 #2394 @OttaviaE
Hey @oliviaguest, I noticed that there is a part of the code that is truncated (nothing important actually). Can I commit the changes in the md file so that the code is not truncated anymore?
