+++ inst/paper.bib
@@ -83,7 +83,7 @@ doi ={10.1177/0146167216684131}
 	title = {DscoreApp: An user-friendly web application for computing the Implicit Association Test D-score},
 	url = {http://fisppa.psy.unipd.it/DscoreApp/},
 	volume = {4},
-	author = {Epifania, O M and Anselmi, Pasquale and Robusto, Egidio},
+	author = {Epifania, Ottavia M and Anselmi, Pasquale and Robusto, Egidio},
 	pages = {1764},
 	date = {2019-10-30},
 	year = {2019},
