### 2020-06-30 26da2e1
Truncated code solved
