### 2020-06-30 #2394 @oliviaguest
Ah, and for the benefit of the other reviewers (and myself):

* https://github.com/OttaviaE/implicitMeasures/issues/3
* https://github.com/OttaviaE/implicitMeasures/issues/2
