### 2020-07-08 OttaviaE/implicitMeasures#2 @wjakethompson
Hi @OttaviaE!

You can get rid of the warnings by using the `.data` pronoun from [{rlang}](https://rlang.r-lib.org/reference/tidyeval-data.html) that is also exported by {ggplot2}, so no need for additional dependencies.

So within the `ggplot()` function, instead of using `data$d` you can use `.data$d`, and this will get rid of the warning. For example:

```r
d_graph <- ggplot(data,
                      aes(y = .data$d, x = .data$d_cres)) +
                      geom_point(col = col_point,
                      size = point_size)
```

If it's possible to make this change, I think it would be helpful. I agree with @eirini-zormpa that the warnings are a little confusing/worrisome when first running the plotting functions, as most users won't understand what is happening since they probably aren't looking at the source code.
