### 2020-07-08 OttaviaE/implicitMeasures#5 @wjakethompson
unit testing for accuracy of d-scores
From the README, it appears that one of the main motivations for this package is that there are often computation errors in the calculation of the D-scores, which raise replicability concerns. I agree that this is a great reason to write a package!

However, when looking at the unit tests for the `computeD()` function, it doesn't look like there are any tests for the accuracy of the calculated values. Right now, the tests check for errors, check the class of return values, and the size of the output (i.e., expected number of rows). These are all important tests, but I think it would be beneficial to also include tests for the accuracy of values. That is, the D-scores returned by the function are what we expect them to be.

This is important to:

1. Ensure the validity of the package
2. Protect against future changes

Presumably, this package will need to be updated at some point in the future as dependencies also get updated. At that time, it would be helpful to have these unit tests in place to ensure that future changes don't mess with any of the calculations, which would then threaten the validity of scores generated from this package.
