### 2020-07-08 #2394 @wjakethompson
I've completed my review and overall I think it looks great! I have two open issues, one that is super minor and is not acceptance blocker (OttaviaE/implicitMeasures#4), and one that I think is a little more important and should be considered prior to publication (OttaviaE/implicitMeasures#5). I'm happy to answer any questions as you look through my suggestions!
