### 2020-07-09 OttaviaE/implicitMeasures#2 @OttaviaE
Thanks @wjakethompson, I'll do that for sure! Thank you for your al suggestions actually, I'll work on them ASAP and provide an appropriate answer under the specific thread =)
