### 2020-07-09 b12c9cf
Plot default settings
