### 2020-07-09 f589164
IAT D-score replicability
