### 2020-07-09 OttaviaE/implicitMeasures#4 @OttaviaE
Thank you for your suggestions! I actually changed a little bit the code, so now there is a consistency in the themes used (it's always `theme_minimal()`. Besides, I also allowed the users to rename the axis names by avoiding the setting `theme(axis.title.x = element_blank())`. I did not leave the deafult ggplot settings because sometimes it might be pretty difficult to use and understand and I wanted users to have pretty graphical representations with the least possible effort.
I hope that with the new changes, the graphs can be easily further modified.
