### 2020-07-09 OttaviaE/implicitMeasures#5 @OttaviaE
I totally agree with what you pointed out. I stored the results from the computation of the IAT and SC-IAT D-score in .rda objects in the `data` folder, and then I used them to test the functions for the computation of the IAT/SC-IAT D-score, and, so far they result in the same exact scores. The new test functions are `test-computeD-replicability` and `test-SCIAT-replicate`.
Thank you for pointing that out, it was really helpful!
