### 2020-07-09 #2394 @OttaviaE
Thanks @wjakethompson, your suggestions were really useful and I think they improved the package. I pushed all the code modifcations and the data sets for the replicability check on Git Hub, so you can have a look at them if you wish. However, I'll wait for the comments of the other reviewer before submitting the package with the new features on CRAN. Thanks again =)
