### 2020-07-23 OttaviaE/implicitMeasures#6 @jflournoy
Some tests fail
Running the tests via devtools, I am getting 2 failures. I'm not sure why--happy to dig deeper if you need.

```
> devtools::test()
Loading implicitMeasures
Testing implicitMeasures
✓ |  OK F W S | Context
✓ |   8       | Dsciat [0.6 s]
✓ |   5       | IATrel_descript_d [1.0 s]
x |   0 2     | SCIAT-replicate [0.7 s]
───────────────────────────────────────────────────────────────────────────────────────
test-SCIAT-replicate.R:32: failure: SC-IAT D-score replicate
dsciat1[, c("participant", "d_sciat")] not equal to Rsciat1[, c("participant", "d_sciat")].
Component "participant": Modes: character, numeric
Component "participant": Attributes: < target is NULL, current is list >
Component "participant": target is character, current is factor

test-SCIAT-replicate.R:34: failure: SC-IAT D-score replicate
dsciat2[, c("participant", "d_sciat")] not equal to Rsciat2[, c("participant", "d_sciat")].
Component "participant": Modes: character, numeric
Component "participant": Attributes: < target is NULL, current is list >
Component "participant": target is character, current is factor
───────────────────────────────────────────────────────────────────────────────────────
✓ |  11       | clean_iat [0.6 s]
✓ |  21       | clean_sciat [0.6 s]
✓ |   7       | computeD-replicability [2.7 s]
✓ |  15       | computeD [5.6 s]
✓ |  10       | d_distr_d_plot [1.7 s]
✓ |  12       | multi_dscore_multi_dsciat [8.4 s]

══ Results ════════════════════════════════════════════════════════════════════════════
Duration: 22.1 s

OK:       89
Failed:   2
Warnings: 0
Skipped:  0
```

And my session info:

```
> sessionInfo()
R version 3.6.3 (2020-02-29)
Platform: x86_64-pc-linux-gnu (64-bit)
Running under: Ubuntu 18.04.4 LTS

Matrix products: default
BLAS:   /usr/lib/x86_64-linux-gnu/blas/libblas.so.3.7.1
LAPACK: /usr/lib/x86_64-linux-gnu/lapack/liblapack.so.3.7.1

locale:
 [1] LC_CTYPE=en_US.UTF-8       LC_NUMERIC=C               LC_TIME=en_US.UTF-8
 [4] LC_COLLATE=en_US.UTF-8     LC_MONETARY=en_US.UTF-8    LC_MESSAGES=en_US.UTF-8
 [7] LC_PAPER=en_US.UTF-8       LC_NAME=C                  LC_ADDRESS=C
[10] LC_TELEPHONE=C             LC_MEASUREMENT=en_US.UTF-8 LC_IDENTIFICATION=C

attached base packages:

other attached packages:
[1] implicitMeasures_0.1.3 testthat_2.3.1

 [1] Rcpp_1.0.4.6      compiler_3.6.3    pillar_1.4.4      prettyunits_1.1.1
 [5] remotes_2.1.1     tools_3.6.3       digest_0.6.25     packrat_0.5.0-25
 [9] pkgbuild_1.0.6    pkgload_1.0.2     memoise_1.1.0     lifecycle_0.2.0
[13] tibble_3.0.1      gtable_0.3.0      pkgconfig_2.0.3   rlang_0.4.6
[17] cli_2.0.2         rstudioapi_0.11   stringr_1.4.0     dplyr_0.8.5
[21] withr_2.2.0       vctrs_0.3.0       desc_1.2.0        fs_1.4.1
[25] devtools_2.2.2    tidyselect_1.1.0  rprojroot_1.3-2   grid_3.6.3
[29] glue_1.4.1        R6_2.4.1          processx_3.4.2    fansi_0.4.1
[33] sessioninfo_1.1.1 tidyr_1.0.2       purrr_0.3.4       callr_3.4.2
[37] ggplot2_3.2.1     magrittr_1.5      backports_1.1.7   scales_1.1.0
[41] ps_1.3.2          ellipsis_0.3.0    usethis_1.6.1     assertthat_0.2.1
[45] xtable_1.8-4      colorspace_1.4-1  stringi_1.4.6     lazyeval_0.2.2
[49] munsell_0.5.0     crayon_1.3.4
```
