### 2020-07-23 OttaviaE/implicitMeasures#7 @jflournoy
Some minor suggestions for language use, missing words.
(Edits in ***bold italics***)

https://github.com/OttaviaE/implicitMeasures/blob/ab6318715872e9075d84dccea1dc516a86059812/inst/paper.Rmd#L33

Could change "Within the past decades..." below to be "In recent decades", "In the past few decades", "Over the last several decades" to be more idiomatically correct.

>Within the past decades, there has been a growing interest in the possibility of assessing people's attitudes, preferences, self-esteem, opinions, and other social-psychological constructs, without directly asking them. This was made possible by the advent of what are called implicit measures. Implicit measures are generally based on the speed and accuracy with which respondents are performing the same categorization task in contrasting conditions.

https://github.com/OttaviaE/implicitMeasures/blob/ab6318715872e9075d84dccea1dc516a86059812/inst/paper.Rmd#L40

>~Despite that~ ***Although*** many R packages exist for computing IAT *D-score* algorithms, no packages exist for scoring the SC-IAT. Additionally, ***the*** majority of existing R packages created for the computation of IAT *D-score* algorithms do not provide all the available algorithms. The packages allowing for the computation of multiple *D-score* algorithms either do not offer the chance to compare their results, or do not disambiguate the specific algorithm they are computing, raising reproducibility issues [@ellithorpe2015].

https://github.com/OttaviaE/implicitMeasures/blob/ab6318715872e9075d84dccea1dc516a86059812/inst/paper.Rmd#L42

>Recently, a Web Application was developed  with `shiny` package [@shiny] for computing the IAT *D-score* [i.e., *DscoreApp*; @dscoreapp]. This app provides an intuitive and easy to use user interface. By giving a detailed explanation of the *D-score* algorithms that can be computed, *DscoreApp* addresses the majority of the above mentioned replicability issues. Moreover, the graphical representation of the results can give an immediate glimpse of the general performance of the respondents. However, *DscoreApp* presents some shortcomings as well. Firstly, since it is a shiny app, it is associated with the most outstanding issue of shiny apps, namely, the replicability of the code, and hence of the results. Specifically, by putting the code into the shiny interface, it is impossible to call it from the command line, and this point is crucial for replication and automation. However, @dscoreapp used a [GitHub](https://github.com/OttaviaE/DscoreApp) repository to let public access the code used for the computation. Despite ***the fact that*** the graphical representations of the results provided by *DscoreApp* are really useful for getting a first idea of the IAT results and ***that*** they are all downloadable in a .pdf format, they cannot be further customized by the users. Moreover, *DscoreApp* computes the *D-score* only for the IAT.
