### 2020-07-23 OttaviaE/implicitMeasures#8 @jflournoy
Function naming conventions
One suggestion to ensure easier or more intuitive usage, especially if this package gains more functions as it is developed, is to establish more consistent internal naming conventions.

Currently the functions are:

https://github.com/OttaviaE/implicitMeasures/blob/ab6318715872e9075d84dccea1dc516a86059812/README.Rmd#L29-L40

For example, you could always use `_` to separate words in function, and functions that do much the same thing could be named very similarly. So for example, `computeD()` and `Dsciat()` could be named `compute_D_iat()` and `compute_D_sciat()`. Capitalization might also be worth making consistent (so always lower-case `d`).

Another suggestion would be to name the plotting functions somewhat like the [`bayesplot`](https://cran.r-project.org/web/packages/bayesplot/bayesplot.pdf) pacakge. For example `d_distr()` might be better named `d_dens()` or  `d_density()`, and `d_plot()` could be `d_scatter()` or `d_point()`. The latter two have the advantage of being congruent with the underlying `ggplot2` geoms.
