### 2020-07-23 #2394 @jflournoy
Please pardon my delay and not seeing the above messages! I've opened three new issues. The first is in regard to the unit tests, two of which fail on my machine. I'm happy to dig deeper if this is not reproducible on your end, @OttaviaE.

https://github.com/OttaviaE/implicitMeasures/issues/6

I'm not sure if this is the right way to do this, but here are a few recommendations for the language in the paper:
https://github.com/OttaviaE/implicitMeasures/issues/7

And finally, some minor suggestions on function naming that may help users remember the names of and find the functions they're looking for:

https://github.com/OttaviaE/implicitMeasures/issues/8
