### 2020-07-24 OttaviaE/implicitMeasures#6 @OttaviaE
Hey! I ran again with `devtools::test()` the tests and I get no failures:

`> devtools::test()
Loading implicitMeasures
Testing implicitMeasures
√ |  OK F W S | Context
√ |   8       | Dsciat [0.7 s]
√ |   5       | IATrel_descript_d [1.1 s]
√ |   2       | SCIAT-replicate [0.8 s]
√ |  11       | clean_iat [0.9 s]
√ |  21       | clean_sciat [1.0 s]
√ |   7       | computeD-replicability [3.5 s]
√ |  15       | computeD [6.4 s]
√ |  10       | d_distr_d_plot [2.1 s]
√ |  12       | multi_dscore_multi_dsciat [9.7 s]

== Results =====================================================================
Duration: 26.3 s

OK:       91
Failed:   0
Warnings: 0
Skipped:  0`

My session info:

`> sessionInfo()
R version 4.0.2 (2020-06-22)
Platform: x86_64-w64-mingw32/x64 (64-bit)
Running under: Windows 10 x64 (build 18362)

Matrix products: default

locale:
[1] LC_COLLATE=Italian_Italy.1252  LC_CTYPE=Italian_Italy.1252    LC_MONETARY=Italian_Italy.1252 LC_NUMERIC=C
[5] LC_TIME=Italian_Italy.1252

attached base packages:

other attached packages:
[1] implicitMeasures_0.1.3 testthat_2.3.2

 [1] Rcpp_1.0.4.6      pillar_1.4.4      compiler_4.0.2    prettyunits_1.1.1 remotes_2.1.1     tools_4.0.2       digest_0.6.25
 [8] pkgbuild_1.0.8    pkgload_1.1.0     gtable_0.3.0      memoise_1.1.0     lifecycle_0.2.0   tibble_3.0.1      pkgconfig_2.0.3
[15] rlang_0.4.6       cli_2.0.2         rstudioapi_0.11   xopen_1.0.0       xfun_0.15         xml2_1.3.2        knitr_1.29
[22] roxygen2_7.1.1    stringr_1.4.0     withr_2.2.0       dplyr_1.0.0       generics_0.0.2    desc_1.2.0        vctrs_0.3.1
[29] fs_1.4.2          devtools_2.3.0    grid_4.0.2        rprojroot_1.3-2   tidyselect_1.1.0  glue_1.4.1        R6_2.4.1
[36] processx_3.4.3    fansi_0.4.1       rcmdcheck_1.3.3   sessioninfo_1.1.1 ggplot2_3.3.2     tidyr_1.1.0       callr_3.4.3
[43] purrr_0.3.4       magrittr_1.5      scales_1.1.1      backports_1.1.8   ps_1.3.3          ellipsis_0.3.1    usethis_1.6.1
[50] assertthat_0.2.1  xtable_1.8-4      colorspace_1.4-1  stringi_1.4.6     munsell_0.5.0     crayon_1.3.4`

It appears I'm using the most updated R version (4.0.2) while you're using an older one. Besides, we're using different systems, but this should not be a problem since I already checked the package functioning on Ubuntu for the CRAN submission. It also appears I'm using the most updated versions of some packages (e.g., `compiler`). Could you please update your R version, run again the tests and let me know whether they still fail? Thank you!
