+++ inst/paper.Rmd
@@ -30,16 +30,16 @@ affiliations:
 
 # Statement of need
 
-Within the past decades, there has been a growing interest in the possibility of assessing people's attitudes, preferences, self-esteem, opinions, and other social-psychological constructs, without directly asking them. This was made possible by the advent of what are called implicit measures. Implicit measures are generally based on the speed and accuracy with which respondents are performing the same categorization task in contrasting conditions. 
+In the past few decades, there has been a growing interest in the possibility of assessing people's attitudes, preferences, self-esteem, opinions, and other social-psychological constructs, without directly asking them. This was made possible by the advent of what are called implicit measures. Implicit measures are generally based on the speed and accuracy with which respondents are performing the same categorization task in contrasting conditions. 
 The assumption underlying their functioning is that respondents' performance will be faster and more accurate in the condition that is consistent with their attitudes, opinions, or preferences.
 The construct of interest is inferred from the difference in the response times between the associative conditions.
 
 Among implicit measures, the Implicit Association Test [IAT; @Greenwald1998] and the Single Category IAT [SC-IAT; @karpinski2006] are the mostly common used ones [@review].
  Both tests result in a differential score (the so-called *D-score*) expressing respondents’ bias in performing the categorization task between conditions. While the scoring of the SC-IAT is based on one single algorithm [@karpinski2006], six different algorithms are available for computing the IAT *D-score* [@Greenwald2003]. The core procedure for the computation of the IAT *D-score* is the same for all the algorithms, which differentiate themselves according to the treatment for extreme fast responses and for the replacement of error responses. 
 
-Despite that many R packages exist for computing IAT *D-score* algorithms, no packages exist for scoring the SC-IAT. Additionally, majority of existing R packages created for the computation of IAT *D-score* algorithms do not provide all the available algorithms. The packages allowing for the computation of multiple *D-score* algorithms either do not offer the chance to compare their results, or do not disambiguate the specific algorithm they are computing, raising reproducibility issues [@ellithorpe2015]. 
+Although many R packages exist for computing IAT *D-score* algorithms, no packages exist for scoring the SC-IAT. Additionally, the majority of existing R packages created for the computation of IAT *D-score* algorithms do not provide all the available algorithms. The packages allowing for the computation of multiple *D-score* algorithms either do not offer the chance to compare their results, or do not disambiguate the specific algorithm they are computing, raising reproducibility issues [@ellithorpe2015]. 
 
-Recently, a Web Application was developed  with `shiny` package [@shiny] for computing the IAT *D-score* [i.e., *DscoreApp*; @dscoreapp]. This app provides an intuitive and easy to use user interface. By giving a detailed explanation of the *D-score* algorithms that can be computed, *DscoreApp* addresses the majority of the above mentioned replicability issues. Moreover, the graphical representation of the results can give an immediate glimpse of the general performance of the respondents. However, *DscoreApp* presents some shortcomings as well. Firstly, since it is a shiny app, it is associated with the most outstanding issue of shiny apps, namely, the replicability of the code, and hence of the results. Specifically, by putting the code into the shiny interface, it is impossible to call it from the command line, and this point is crucial for replication and automation. However, @dscoreapp used a [GitHub](https://github.com/OttaviaE/DscoreApp) repository to let public access the code used for the computation. Despite the graphical representations of the results provided by *DscoreApp* are really useful for getting a first idea of the IAT results and they are all downloadable in a .pdf format, they cannot be further customized by the users. Moreover, *DscoreApp* computes the *D-score* only for the IAT. 
[…]
