### 2020-07-24 42fd84c
Functions names consistency
