+++ R/d_density.R
@@ -1,0 +1,158 @@
+#' Plot IAT or SC-IAT scores (distribution)
+#'
+#' Plot the distribution of the IAT \emph{D-score} or the SC-IAT \emph{D}.
+#'
+#' @param data Dataframe with either class \code{dscore} or \code{dsciat}.
+#' @param graph String. Indicates the graphs to display. Default is
+#'                \code{histogram}
+#' @param n_bin Numeric. Indicates the number of bins to display.
+#' @param col_fill String. Indicates the color for filling the bars of the
+#'                  \code{histogram} or the curve of the \code{density}. Default
+#'                  is \code{royalblue}.
+#' @param col_point String. Indicates the color for the individual scores --each
+#'                   point -- in the violin plot. Default is \code{red}.
+#' @param include_stats Logical. Indicates whether to add descriptive statistics.
+#'                       The \code{mean} is depicted with a solid line. The two
+#'                       dashed lines represent +/- 2 \emph{s.d.} from the mean.
+#'                       Default is \code{FALSE}.
+#'
+#' @return A ggplot object.
+#'
+#' @export
+#'
+#' @import ggplot2
+#'
+#' @examples
+#' \donttest{
+#'   # Plotting the IAT D-score
+#'   data("raw_data") # import data
+#'   iat_cleandata <- clean_iat(raw_data, sbj_id = "Participant",
+#'                           block_id = "blockcode",
+#'                           mapA_practice = "practice.iat.Milkbad",
+#'                           mapA_test = "test.iat.Milkbad",
+#'                           mapB_practice = "practice.iat.Milkgood",
+#'                           mapB_test = "test.iat.Milkgood",
+#'                           latency_id = "latency",
+#'                           accuracy_id = "correct",
+#'                           trial_id = "trialcode",
+#'                           trial_eliminate = c("reminder", "reminder1"),
+#'                           demo_id = "blockcode",
+#'                           trial_demo = "demo")
+#'   iat_data <- iat_cleandata[[1]]
+#' # calculate D-score
+#'   iat_dscore <- computeD(iat_data,
+#'                        Dscore =  "d2")
+#'   d_density(iat_dscore) # Default graph
+#'   d_density(iat_dscore, graph = "histogram",
+#'           n_bin = 30) # Histogram with a different number of bins
+#'   d_density(iat_dscore, graph = "density") # IAT D-score density plot
+#'   d_density(iat_dscore, graph = "violin") # IAT D-score violin plot
+#'
+#'   # Plot the SC-IAT D for the first SC-IAT
+#'   data("raw_data") # load data
+#'   sciat_data <- clean_sciat(raw_data, sbj_id = "Participant",
+#'                          block_id = "blockcode",
+#'                          latency_id = "latency",
+#'                          accuracy_id = "correct",
+#'                          block_sciat_1 = c("test.sc_dark.Darkbad",
+#'                                            "test.sc_dark.Darkgood"),
+#'                          block_sciat_2 = c("test.sc_milk.Milkbad",
+#'                                            "test.sc_milk.Milkgood"),
+#'                          trial_id  = "trialcode",
+#'                          trial_eliminate = c("reminder",
+#'                                              "reminder1"))
+#'
+#'  sciat1 <- sciat_data[[1]] # compute D for the first SC-IAT
+#'  d_sciat1 <- Dsciat(sciat1,
+#'                   mappingA = "test.sc_dark.Darkbad",
+#'                   mappingB = "test.sc_dark.Darkgood",
+#'                   non_response = "alert")
+#'   d_density(d_sciat1, graph = "histogram",
+#'           include_stats = TRUE) # SC-IAT D histogram with descriptive
+#'                                    #  statistics
[…]
