+++ R/iatdscores.R
@@ -1,0 +1,14 @@
+#' Data set with IAT D-scores
+#'
+#' A data set cotaining the results for all the possible D-score algorithms for the
+#' IAT. All the algorithms are identified by their corresponding label (such
+#' as "dscore_d1"). This data set is used for testing the replicability of
+#' the results of the "computeD()" function over time.
+#'
+#' @docType data
+#' @keywords datasets
+#' @name iatdscores
+#' @usage data("iatdscores")
+#' @format A dataframe with 7 variables, the first one contains the respondents'
+#' id, the otehr 6 colums contain a specific D-score algorithm.
+"iatdscores"

[…]
