+++ R/dsciat2.R
@@ -1,0 +1,13 @@
+#' Data set with SC-IAT D-scores (Milk)
+#'
+#' A data set containing the results of the computation of the D-score on the
+#' Dark SC-IAT data set. This data set is used for testing the replicability of the
+#' results obtained with the "Dsciat()" functions.
+#'
+#' @docType data
+#' @keywords datasets
+#' @name dsciat2
+#' @usage data("dsciat2")
+#' @format A dataframe with 15 variables, as those described in the documentation for the
+#' "Dsciat()" function.
+"dsciat2"
