+++ man/dsciat2.Rd
@@ -1,0 +1,19 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/dsciat2.R
+\docType{data}
+\name{dsciat2}
+\alias{dsciat2}
+\title{Data set with SC-IAT D-scores (Milk)}
+\format{
+A dataframe with 15 variables, as those described in the documentation for the
+"Dsciat()" function.
+}
+\usage{
+data("dsciat2")
+}
+\description{
+A data set containing the results of the computation of the D-score on the
+Dark SC-IAT data set. This data set is used for testing the replicability of the
+results obtained with the "Dsciat()" functions.
+}
+\keyword{datasets}
