+++ man/iatdscores.Rd
@@ -1,0 +1,20 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/iatdscores.R
+\docType{data}
+\name{iatdscores}
+\alias{iatdscores}
+\title{Data set with IAT D-scores}
+\format{
+A dataframe with 7 variables, the first one contains the respondents'
+id, the otehr 6 colums contain a specific D-score algorithm.
+}
+\usage{
+data("iatdscores")
+}
+\description{
+A data set containing the results for all the possible D-score algorithms for the
+IAT. All the algorithms are identified by their corresponding label (such
+as "dscore_d1"). This data set is used for testing the replicability of
+the results of the "computeD()" function over time.
+}
+\keyword{datasets}
