### 2020-07-24 ffc09d8
New functions names
