### 2020-07-24 OttaviaE/implicitMeasures#8 @OttaviaE
@jflournoy You are absolutely right. I changed some of the functions names to make them more consistent. Clearly, the package version with the updated names is not currently on CRAN (I plan to submit it by the end of the weekend) but I already pushed the changes on GitHub.
