### 2020-07-24 OttaviaE/implicitMeasures#6 @OttaviaE
@jflournoy Since I changed the names of the functions, I re-ran the tests with `devtools::test()`, and I still get no failures (session info are the same as those in the previous comment):

`> test()
Loading implicitMeasures
Testing implicitMeasures
√ |  OK F W S | Context
√ |   5       | IAT_rel_descript_d [1.3 s]
√ |  11       | clean_iat [0.9 s]
√ |  21       | clean_sciat [1.0 s]
√ |   7       | compute_iat-replicate [3.1 s]
√ |  15       | compute_iat [6.4 s]
√ |   2       | compute_sciat-replicate [0.9 s]
√ |   8       | compute_sciat [0.6 s]
√ |  10       | d_density_d_plot [2.1 s]
√ |  12       | multi_dscore_multi_dsciat [9.8 s]

== Results =====================================================================
Duration: 26.1 s

OK:       91
Failed:   0
Warnings: 0
Skipped:  0
> test()
Loading implicitMeasures
Testing implicitMeasures
√ |  OK F W S | Context
√ |   5       | IAT_rel_descript_d [1.2 s]
√ |  11       | clean_iat [1.0 s]
√ |  21       | clean_sciat [1.0 s]
√ |   7       | compute_iat-replicate [2.9 s]
√ |  15       | compute_iat [5.8 s]
√ |   2       | compute_sciat-replicate [0.7 s]
√ |   8       | compute_sciat [0.6 s]
√ |  10       | d_density_d_plot [2.1 s]
√ |  12       | multi_dscore_multi_dsciat [9.0 s]

== Results =====================================================================
Duration: 24.3 s

OK:       91
Failed:   0
Warnings: 0
Skipped:  0`

When I run `devtools::check()`, I get no red flags as well.
