### 2020-07-24 13f72a1
Typos + new functions names
