### 2020-07-24 c90dcad
Updating functions names
