+++ DESCRIPTION
@@ -1,7 +1,7 @@
 Package: implicitMeasures
 Type: Package
-Title: Computes the Scores for Different Implicit Measures
-Version: 0.1.3
+Title: Compute Scores for Different Implicit Measures
+Version: 0.2.0
 Author: Ottavia M. Epifania [aut, cre],
     Pasquale Anselmi [ctb], 
     Egidio Robusto [ctb]
