### 2020-07-24 OttaviaE/implicitMeasures#6 @jflournoy
Thank you for checking that. Perhaps it would be good to set required R
version and package versions. I'll try to locate the cause of the test
failure, but if it is a version issue it would be good to impose those on
the dependencies to ensure that the tests pass on the versions allowed by
the package.
