### 2020-07-24 OttaviaE/implicitMeasures#6 @OttaviaE
Thank you, I appreaciate that.
