### 2020-07-26 664b572
Forcing a variable to be a character
