+++ inst/paper.bib
@@ -35,7 +35,7 @@ year = {2002}
 author = {Greenwald, Anthony G and McGhee, Debbie E and Schwartz, Jordan L K},
 doi = {10.1037/0022-3514.74.6.1464},
 issn = {0022-3514},
-journal = {Journal of Personality and Soclal Psychology},
+journal = {Journal of Personality and Social Psychology},
 number = {6},
 pages = {1464--1480},
 pmid = {9654756},
