+++ R/dsciat2.R
@@ -2,12 +2,12 @@
 #'
 #' A data set containing the results of the computation of the D-score on the
 #' Dark SC-IAT data set. This data set is used for testing the replicability of the
-#' results obtained with the "Dsciat()" functions.
+#' results obtained with the \code{compute_sciat()} functions.
 #'
 #' @docType data
 #' @keywords datasets
 #' @name dsciat2
 #' @usage data("dsciat2")
 #' @format A dataframe with 15 variables, as those described in the documentation for the
-#' "Dsciat()" function.
+#' \code{compute_sciat()} function.
 "dsciat2"
