+++ R/iatdscores.R
@@ -3,12 +3,12 @@
 #' A data set containing the results for all the possible D-score algorithms for the
 #' IAT. All the algorithms are identified by their corresponding label (such
 #' as "dscore_d1"). This data set is used for testing the replicability of
-#' the results of the "computeD()" function over time.
+#' the results of the \code{compute_iat()} function over time.
 #'
 #' @docType data
 #' @keywords datasets
 #' @name iatdscores
 #' @usage data("iatdscores")
 #' @format A dataframe with 7 variables, the first one contains the respondents'
-#' id, the otehr 6 colums contain a specific D-score algorithm.
+#' id, the other 6 columns contain a specific D-score algorithm.
 "iatdscores"
