### 2020-07-27 OttaviaE/implicitMeasures#6 @OttaviaE
Hey @jflournoy I got it, I solved it! There was a problem specifically with Ubuntu. I used `devtools::check_r_hub()` to check whether it's working on every platform, and it appears it does. Let me know whether it's working for you again, and thanks for your suggestions.
I'll wait for your okay before submitting to CRAN again. Cheers!
