### 2020-07-27 OttaviaE/implicitMeasures#6 @jflournoy
Awesome! I reran the checks and they're all okay now.
