### 2020-07-27 OttaviaE/implicitMeasures#6 @OttaviaE
Sweet! Thanks again!
I'm running the last checks and then I'll submit to CRAN =)
