### 2020-07-27 #2394 @jflournoy
Okay, @oliviaguest, I believe all boxes are checked. Thank you so much, @OttaviaE, for being so responsive; I'm glad my suggestions were harmonious with your goals for this software and the paper. And more so, thanks for writing this package!
