### 2020-07-27 59c0f3e
CRAN comments
