+++ cran-comments.md
@@ -1,12 +1,10 @@
 ## Test environments
-* local OS X install, R 3.6.3
-* ubuntu 14.04 (on travis-ci), R 3.6.3
-* win-builder (devel and release)
+* local R installation, R 4.0.2
+* ubuntu 16.04 (on travis-ci), R 4.0.2
+* win-builder (devel)
 
 ## R CMD check results
 
-0 errors | 0 warnings | 1 note
+0 errors | 0 warnings | 0 note
 
-A bibliography was added to the README file (file "paper.bib"). ALl vignettes have been corrected to make them more clear and understandable. 
-A new folder ("joss") was added at the top level. The folder contains a paper about the package that is going to be submitted to the Journal of Open Source Software. 
-The presence of these two new files at the top level produces the note.
+New versions of the package with updeated functions names
