### 2020-07-30 #2394 @oliviaguest
Oh, perfect! @OttaviaE Can you make the title of the zenodo repo be exactly: ` Implicit measures with reproducible results: The implicitMeasures package`, please?
