### 2020-08-01 #2394 @kthyng
Hi @OttaviaE! I see the title of your package at Zenodo is the same as your paper title – excellent – but can you also have the author list match your paper?

Also, can I verify the version that you want associated with your JOSS submission? It looks to me like it should be v0.2.0?
