### 2020-08-01 OttaviaE/implicitMeasures#7 @kthyng
@OttaviaE and @jflournoy What is the status of this issue, in regard to finishing the JOSS publication?
