### 2020-08-01 #2394 @kthyng
I'll just note here that there is an issue that is still [open](https://github.com/OttaviaE/implicitMeasures/issues/7) that appears to be associated with this review that I am waiting to hear back about.
