### 2020-08-01 ed2db76
Some changes to improve capitalization in refs
For JOSS review [#2394](https://github.com/openjournals/joss-reviews/issues/2394)
