+++ inst/paper.bib
@@ -80,7 +80,8 @@ doi ={10.1177/0146167216684131}
 	issn = {2475-9066},
 	number = {42},
 	publisher = {The Open Journal},
-	title = {DscoreApp: An user-friendly web application for computing the Implicit Association Test D-score},
+	title = {DscoreApp: An user-friendly 
+	for computing the {Implicit Association Test D-score}},
 	url = {http://fisppa.psy.unipd.it/DscoreApp/},
 	volume = {4},
 	author = {Epifania, O M and Anselmi, Pasquale and Robusto, Egidio},
@@ -91,7 +92,7 @@ doi ={10.1177/0146167216684131}
 	day = {30},
 }
  @Manual{shiny,
-    title = {shiny: Web Application Framework for R},
+    title = {shiny: Web Application Framework for {R}},
     author = {Winston Chang and Joe Cheng and JJ Allaire and Yihui Xie and Jonathan McPherson},
     year = {2020},
     note = {R package version 1.4.0.1},
@@ -117,7 +118,7 @@ doi = {10.31124/advance.11914416.v1}
 }
 @Book{ggplot2,
     author = {Hadley Wickham},
-    title = {ggplot2: Elegant Graphics for Data Analysis},
+    title = {{ggplot2: Elegant Graphics for Data Analysis}},
     publisher = {Springer-Verlag New York},
     year = {2016},
     isbn = {978-3-319-24277-4},
