### 2020-08-01 OttaviaE/implicitMeasures#7 @OttaviaE
Hi @kthyng! I actually changed the paper according to @jflournoy's suggestions, I don't know why the issue is still open
