### 2020-08-01 #2394 @OttaviaE
Hey @kthyng! I added my co-authors on Zenodo, but I also left @oliviaguest in the author list.

Regarding the version, yes, the version associated to my joss submission is version v0.2.0
