### 2020-08-01 OttaviaE/implicitMeasures#7 @kthyng
Ok great, hopefully we can hear from @jflournoy that this is finished and he can close it. We should wait until at least a few business days have passed to give a chance.
