### 2020-08-03 #2394 @danielskatz
@OttaviaE -  can you merge the small changes in https://github.com/OttaviaE/implicitMeasures/pull/10 ?  Then we can finish the acceptance
