### 2020-08-03 #2394 @danielskatz
@openjournals/dev  - while the DOI is resolving, the PDF does not seem to load - can you check?
