### 2020-08-06 #2394 @OttaviaE
@danielskatz thank you!

And also, thank you @eirini-zormpa, @jflournoy, and @wjakethompson for your kind and useful comments, you all definetly helped in improving my package =)
Of course, thank you @oliviaguest for being such a great editor!
